Initial Endpoint for Metrics API This will serve metrics data to external services Metrics API consists of a Function App with an Azure HTTPTrigger event function for each endpoint.
The first endpoint added here is the fetch all Participant Uploads from the metrics database.

For now, query limit is locked at 50 records and returns 1 page of results

Piipan.Metrics.Models were extracted to be shared between the API and other metrics subsystems.

PiipanMetricsApi contains a few Serializers to help format the http responses

For now, the API is using the openapi standard as a reference only